+++ lib/chef/http/simple.rb
@@ -11,6 +11,11 @@ class Chef
       use Decompressor
       use CookieManager
 
+      # ValidateContentLength should come after Decompressor
+      # because the order of middlewares is reversed when handling
+      # responses.
+      use ValidateContentLength
+
     end
   end
 end